separate state store interface from implementation
